Split up chat ts files
